make a utils module
